pyproject.toml
[tool.poetry]
package-mode = false

[project]
name = "agentic-review-platform"
version = "0.1.0"
description = "Agentic Engineering Review Platform"
authors = [
    {name = "Ravi Yadavilli", email = "ryadavillicloud@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = ["fastapi (>=0.135.1,<0.136.0)", "uvicorn[standard] (>=0.42.0,<0.43.0)", "openai (>=2.29.0,<3.0.0)", "pydantic (>=2.12.5,<3.0.0)", "python-dotenv (>=1.2.2,<2.0.0)", "httpx (>=0.28.1,<0.29.0)", "mcp (>=1.25,<2)", "opentelemetry-sdk (>=1.40.0,<2.0.0)", "opentelemetry-api (>=1.40.0,<2.0.0)", "opentelemetry-instrumentation-fastapi (>=0.61b0,<0.62)", "opentelemetry-instrumentation-httpx (>=0.61b0,<0.62)", "azure-monitor-opentelemetry-exporter (>=1.0.0b49,<2.0.0)"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "pytest-cov (>=7.1.0,<8.0.0)",
    "ruff (>=0.15.7,<0.16.0)",
    "mypy (>=1.19.1,<2.0.0)",
    "pre-commit (>=4.5.1,<5.0.0)"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "scenario: End to end scenario tests",
    "integration: mark test as integration test",
    "unit: mark test as unit test",
    "slow: mark test as slow (e.g., requires API calls)"
]
asyncio_mode = "auto"
addopts = "--cov=app --cov-branch --cov-report=term-missing"

[tool.coverage.run]
source = ["app"]
branch = true
omit = [
    "app/**/__init__.py"
]

[tool.coverage.report]
show_missing = true
fail_under = 0

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
no_implicit_optional = true
strict_equality = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = ["uvicorn.*", "fastapi.*", "openai.*", "httpx.*", "pydantic.*", "mcp.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "opentelemetry.instrumentation.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "azure.*"
ignore_missing_imports = true

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "N", "S"]
ignore = ["S101"]  

[tool.ruff.lint.isort]
known-first-party = ["app"]

